Effectful.PosixCompat.Dynamic

Description

Provides a dynamic effect for System.PosixCompat.Files.

Since: 0.1

Effect

data PosixCompatDynamic :: Effect where

Dynamic effect for System.PosixCompat.Files.

Since: 0.1

Constructors

SetFileMode :: FilePath -> FileMode -> PosixCompatDynamic m ()

SetFdMode :: Fd -> FileMode -> PosixCompatDynamic m ()

SetFileCreationMask :: FileMode -> PosixCompatDynamic m FileMode

FileAccess :: FilePath -> Bool -> Bool -> Bool -> PosixCompatDynamic m Bool

FileExist :: FilePath -> PosixCompatDynamic m Bool

GetFileStatus :: FilePath -> PosixCompatDynamic m FileStatus

GetFdStatus :: Fd -> PosixCompatDynamic m FileStatus

GetSymbolicLinkStatus :: FilePath -> PosixCompatDynamic m FileStatus

CreateNamedPipe :: FilePath -> FileMode -> PosixCompatDynamic m ()

CreateDevice :: FilePath -> FileMode -> DeviceID -> PosixCompatDynamic m ()

CreateLink :: FilePath -> FilePath -> PosixCompatDynamic m ()

RemoveLink :: FilePath -> PosixCompatDynamic m ()

CreateSymbolicLink :: FilePath -> FilePath -> PosixCompatDynamic m ()

ReadSymbolicLink :: FilePath -> PosixCompatDynamic m FilePath

Rename :: FilePath -> FilePath -> PosixCompatDynamic m ()

SetOwnerAndGroup :: FilePath -> UserID -> GroupID -> PosixCompatDynamic m ()

SetFdOwnerAndGroup :: Fd -> UserID -> GroupID -> PosixCompatDynamic m ()

SetSymbolicLinkOwnerAndGroup :: FilePath -> UserID -> GroupID -> PosixCompatDynamic m ()

SetFileTimes :: FilePath -> EpochTime -> EpochTime -> PosixCompatDynamic m ()

TouchFile :: FilePath -> PosixCompatDynamic m ()

SetFileSize :: FilePath -> FileOffset -> PosixCompatDynamic m ()

SetFdSize :: Fd -> FileOffset -> PosixCompatDynamic m ()

GetPathVar :: FilePath -> PathVar -> PosixCompatDynamic m Limit

GetFdPathVar :: Fd -> PathVar -> PosixCompatDynamic m Limit

setFileMode :: PosixCompatDynamic :> es => FilePath -> FileMode -> Eff es ()

Lifted setFileMode.

Since: 0.1

setFdMode :: PosixCompatDynamic :> es => Fd -> FileMode -> Eff es ()

Lifted setFdMode.

Since: 0.1

setFileCreationMask :: PosixCompatDynamic :> es => FileMode -> Eff es FileMode

Lifted setFileCreationMask.

Since: 0.1

fileAccess :: PosixCompatDynamic :> es => FilePath -> Bool -> Bool -> Bool -> Eff es Bool

Lifted fileAccess.

Since: 0.1

fileExist :: PosixCompatDynamic :> es => FilePath -> Eff es Bool

Lifted fileExist.

Since: 0.1

getFileStatus :: PosixCompatDynamic :> es => FilePath -> Eff es FileStatus

Lifted getFileStatus.

Since: 0.1

getFdStatus :: PosixCompatDynamic :> es => Fd -> Eff es FileStatus

Lifted getFdStatus.

Since: 0.1

getSymbolicLinkStatus :: PosixCompatDynamic :> es => FilePath -> Eff es FileStatus

Lifted getSymbolicLinkStatus.

Since: 0.1

createNamedPipe :: PosixCompatDynamic :> es => FilePath -> FileMode -> Eff es ()

Lifted createNamedPipe.

Since: 0.1

createDevice :: PosixCompatDynamic :> es => FilePath -> FileMode -> DeviceID -> Eff es ()

Lifted createDevice.

Since: 0.1

createLink :: PosixCompatDynamic :> es => FilePath -> FilePath -> Eff es ()

Lifted createLink.

Since: 0.1

removeLink :: PosixCompatDynamic :> es => FilePath -> Eff es ()

Lifted removeLink.

Since: 0.1

createSymbolicLink :: PosixCompatDynamic :> es => FilePath -> FilePath -> Eff es ()

Lifted createSymbolicLink.

Since: 0.1

readSymbolicLink :: PosixCompatDynamic :> es => FilePath -> Eff es FilePath

Lifted readSymbolicLink.

Since: 0.1

rename :: PosixCompatDynamic :> es => FilePath -> FilePath -> Eff es ()

Lifted rename.

Since: 0.1

setOwnerAndGroup :: PosixCompatDynamic :> es => FilePath -> UserID -> GroupID -> Eff es ()

Lifted setOwnerAndGroup.

Since: 0.1

setFdOwnerAndGroup :: PosixCompatDynamic :> es => Fd -> UserID -> GroupID -> Eff es ()

Lifted setFdOwnerAndGroup.

Since: 0.1

setSymbolicLinkOwnerAndGroup :: PosixCompatDynamic :> es => FilePath -> UserID -> GroupID -> Eff es ()

Lifted setSymbolicLinkOwnerAndGroup.

Since: 0.1

setFileTimes :: PosixCompatDynamic :> es => FilePath -> EpochTime -> EpochTime -> Eff es ()

Lifted setFileTimes.

Since: 0.1

touchFile :: PosixCompatDynamic :> es => FilePath -> Eff es ()

Lifted touchFile.

Since: 0.1

setFileSize :: PosixCompatDynamic :> es => FilePath -> FileOffset -> Eff es ()

Lifted setFileSize.

Since: 0.1

setFdSize :: PosixCompatDynamic :> es => Fd -> FileOffset -> Eff es ()

Lifted setFdSize.

Since: 0.1

getPathVar :: PosixCompatDynamic :> es => FilePath -> PathVar -> Eff es Limit

Lifted getPathVar.

Since: 0.1

getFdPathVar :: PosixCompatDynamic :> es => Fd -> PathVar -> Eff es Limit

Lifted getFdPathVar.

Since: 0.1

Handler

runPosixCompatDynamicIO :: IOE :> es => Eff (PosixCompatDynamic : es) a -> Eff es a

Runs PosixCompatDynamic in IO.

Since: 0.1

PathType

data PathType

Path type. @since 0.1

Constructors

PathTypeFile
Since: 0.1

PathTypeDirectory
Since: 0.1

PathTypeSymbolicLink
Since: 0.1

PathTypeOther
Since: 0.1

displayPathType :: IsString a => PathType -> a

String representation of PathType.

Since: 0.1

Functions

throwIfWrongPathType :: PosixCompatDynamic :> es => String -> PathType -> FilePath -> Eff es ()

Throws IOException if the path does not exist or the expected path type does not match actual.

Since: 0.1

isPathType :: PosixCompatDynamic :> es => PathType -> FilePath -> Eff es Bool

Checks that the path type matches the expectation. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

getPathType :: PosixCompatDynamic :> es => FilePath -> Eff es PathType

Returns the type for a given path without following symlinks. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

Optics

_PathTypeFile :: Prism' PathType ()

Since: 0.1

_PathTypeDirectory :: Prism' PathType ()

Since: 0.1

_PathTypeSymbolicLink :: Prism' PathType ()

Since: 0.1

_PathTypeOther :: Prism' PathType ()

Since: 0.1

Utils

throwPathIOError

Arguments

:: (HasCallStack, MonadThrow m)

=> FilePath
Path upon which the IO operation failed.

-> String
String location (e.g. function name).

-> IOErrorType
Type of exception.

-> String
Description.

-> m a

Helper for throwing IOException.

Since: 0.1